requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
mcp>=1.0.0
openai>=1.0.0
